refactor(rust): split mime_utils en attachments/dto (≤300 LOC)
