+++ src/bin/email_api_dir/mailbox/mime_attachments.rs
@@ -1,0 +1,101 @@
+// extracted from mime_utils cycle 16
+use super::super::super::*;
+#[allow(unused_imports)]
+use base64::Engine as _;
+use chrono::Utc;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+use super::mime_utils::*;
+
+#[derive(Clone)]
+pub(crate) struct ExtractedAttachment {
+    pub id: String,
+    pub filename: String,
+    pub content_type: String,
+    pub size: u64,
+    pub kind: String,
+    pub data: Vec<u8>,
+}
+
+fn walk_mime_attachments(
+    part: &mailparse::ParsedMail<'_>,
+    out: &mut Vec<ExtractedAttachment>,
+    index: &mut usize,
+) {
+    if !part.subparts.is_empty() {
+        for sub in &part.subparts {
+            walk_mime_attachments(sub, out, index);
+        }
+        return;
+    }
+
+    let content_type = part.ctype.mimetype.to_ascii_lowercase();
+    let disp = part.get_content_disposition();
+    let disp_kind = format!("{:?}", disp.disposition).to_ascii_lowercase();
+    let filename = disp
+        .params
+        .get("filename")
+        .cloned()
+        .or_else(|| part.ctype.params.get("name").cloned())
+        .filter(|s| !s.trim().is_empty())
+        .unwrap_or_else(|| format!("attachment-{}", *index + 1));
+
+    let is_attachment = disp_kind == "attachment"
+        || disp.params.contains_key("filename")
+        || part.ctype.params.contains_key("name")
+        || (!content_type.starts_with("text/") && content_type != "application/pgp-signature");
+
+    if !is_attachment {
+        return;
+    }
+
+    let bytes = part.get_body_raw().unwrap_or_default();
+    let kind = infer_attachment_kind(&content_type, &filename).to_string();
+    let id = format!("att-{}", *index);
+    *index += 1;
+
+    out.push(ExtractedAttachment {
+        id,
+        filename,
+        content_type,
+        size: bytes.len() as u64,
+        kind,
+        data: bytes,
+    });
+}
+
+pub(crate) fn extract_attachments_for_ui(email: &Email) -> Vec<ExtractedAttachment> {
+    let raw_mime = raw_mime_from_email(email);
+    if let Ok(parsed) = mailparse::parse_mail(raw_mime.as_bytes()) {
+        let mut out = Vec::new();
+        let mut idx = 0usize;
+        walk_mime_attachments(&parsed, &mut out, &mut idx);
+        return out;
+    }
+
+    if looks_like_raw_multipart_dump(&email.body) {
+        let first_line = match email.body.lines().next() {
+            Some(l) => l.trim(),
+            None => return Vec::new(),
+        };
+        let boundary = first_line
+            .trim_start_matches("--")
+            .trim_end_matches("--")
+            .trim();
+        if boundary.is_empty() {
+            return Vec::new();
+        }
+        let synthetic = format!(
+            "Content-Type: multipart/mixed; boundary=\"{}\"\r\n\r\n{}",
+            boundary, email.body
+        );
+        if let Ok(parsed) = mailparse::parse_mail(synthetic.as_bytes()) {
+            let mut out = Vec::new();
+            let mut idx = 0usize;
+            walk_mime_attachments(&parsed, &mut out, &mut idx);
+            return out;
+        }
+    }
+
+    Vec::new()
+}
