+++ src/bin/email_api_dir/mailbox/mime_dto.rs
@@ -1,0 +1,154 @@
+// extracted from mime_utils cycle 16
+use super::super::super::*;
+#[allow(unused_imports)]
+use base64::Engine as _;
+use chrono::Utc;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+use super::mime_utils::*;
+
+fn decoded_mail_body_for_ui(email: &Email) -> (String, String, String) {
+    let raw_mime = raw_mime_from_email(email);
+
+    if let Ok(parsed) = mailparse::parse_mail(raw_mime.as_bytes()) {
+        if let Some(decoded) = decode_parts_from_parsed(&parsed) {
+            return decoded;
+        }
+
+        if let Ok(body) = parsed.get_body() {
+            if looks_like_raw_multipart_dump(&body) {
+                if let Some(decoded) = decode_from_synthetic_boundary(&body) {
+                    return decoded;
+                }
+            }
+
+            let looks_html = body.to_ascii_lowercase().contains("<html")
+                || body.contains("</")
+                || body.contains("<p");
+            if looks_html {
+                let preview = compact_preview(&strip_tags(&body)).chars().take(160).collect();
+                return (body, "html".to_string(), preview);
+            }
+            let preview = compact_preview(&body).chars().take(160).collect();
+            return (body, "text".to_string(), preview);
+        }
+    }
+
+    if let Some(decoded) = decode_from_synthetic_boundary(&email.body) {
+        return decoded;
+    }
+
+    let body_type = if email.body.to_ascii_lowercase().contains("<html")
+        || email.body.contains("</")
+        || email.body.contains("<p")
+    {
+        "html"
+    } else {
+        "text"
+    };
+    let plain = if body_type == "html" {
+        strip_tags(&email.body)
+    } else {
+        email.body.clone()
+    };
+    let preview = compact_preview(&plain).chars().take(160).collect();
+    (email.body.clone(), body_type.to_string(), preview)
+}
+
+pub(crate) fn email_to_dto(email: &Email, folder: &str, include_body: bool) -> EmailDto {
+    let flags_l: Vec<String> = email.flags.iter().map(|f| f.to_ascii_lowercase()).collect();
+    let is_read = flags_l.iter().any(|f| f == "seen" || f == "\\seen");
+    let is_starred = flags_l
+        .iter()
+        .any(|f| f == "flagged" || f == "\\flagged" || f == "starred");
+    let (decoded_body, body_type, preview) = decoded_mail_body_for_ui(email);
+    let date = {
+        let ms = email.internal_date.timestamp_millis();
+        chrono::DateTime::from_timestamp_millis(ms)
+            .map(|d| d.to_rfc3339())
+            .unwrap_or_else(|| Utc::now().to_rfc3339())
+    };
+    let message_id = email
+        .headers
+        .iter()
+        .find_map(|(k, v)| {
+            if k.eq_ignore_ascii_case("message-id") && !v.is_empty() {
+                Some(v.clone())
+            } else if k.to_ascii_lowercase().starts_with("message-id:") {
+                let val = k.splitn(2, ':').nth(1).unwrap_or("").trim();
+                if !val.is_empty() {
+                    Some(val.to_string())
+                } else if !v.is_empty() {
+                    Some(v.clone())
+                } else {
+                    None
+                }
+            } else {
+                None
+            }
+        })
+        .filter(|s| !s.is_empty())
+        .unwrap_or_else(|| {
+            if !email.id.is_empty() {
+                email.id.clone()
+            } else if email.uid > 0 {
+                format!("uid-{}", email.uid)
+            } else {
+                Uuid::new_v4().to_string()
+            }
+        });
+    let id = if email.id.is_empty() {
+        message_id
+            .trim_matches(|c| c == '<' || c == '>')
+            .to_string()
+    } else {
+        email.id.clone()
+    };
+    let to_list: Vec<EmailAddressDto> = email
+        .to
+        .split(',')
+        .map(str::trim)
+        .filter(|s| !s.is_empty())
+        .map(parse_address)
+        .collect();
+    let attachments = extract_attachments_for_ui(email);
+    let attachments_json: Vec<serde_json::Value> = attachments
+        .iter()
+        .map(|att| {
+            serde_json::json!({
+                "id": att.id,
+                "filename": att.filename,
+                "contentType": att.content_type,
+                "size": att.size,
+                "type": att.kind,
+                "downloadUrl": format!("/api/emails/{}/attachments/{}", id, att.id),
+            })
+        })
+        .collect();
+
+    EmailDto {
+        id,
+        thread_id: message_id.clone(),
+        folder: folder.to_ascii_lowercase(),
+        from: parse_address(&email.from),
+        to: to_list,
+        subject: email.subject.clone(),
+        preview,
+        body: if include_body {
+            decoded_body.clone()
+        } else {
+            String::new()
+        },
+        body_type,
+        date: date.clone(),
+        received_at: date,
+        is_read,
+        is_starred,
+        is_important: false,
+        has_attachments: !attachments_json.is_empty(),
+        attachments: attachments_json,
+        labels: vec![],
+        size: email.body.len() as u64,
+        message_id,
+    }
+}
